02-Flex-header
Completed 02-Flex-header, took a while and I had to use some help but
eventually got it working when I learned that you have to set the direct
parent as the flex container for it to work correctly, I was trying to
make left-links a flex container and it wasn't working.  Set up the <ul>
with a class and turned that into a flex container instead, which worked.