05-flex-modal
Managed to make this assignment look like the desired-outcome.png file,
not sure if I did it in the right way however it looks the same.  I was
confused about how to have the "header" class align with the text and
continue/cancel buttons vertically (column) as well as icon and cross
button horizontally (row).